+++ packages/codec/lib/core.ts
@@ -4,7 +4,10 @@ const debug = debugModule("codec:core");
 import * as Ast from "@truffle/codec/ast";
 import * as Abi from "@truffle/codec/abi";
 import * as Pointer from "@truffle/codec/pointer";
-import * as Allocation from "@truffle/codec/allocate/types";
+import {
+  EventAllocation,
+  CalldataAllocation
+} from "@truffle/codec/allocate/abi/types";
 import {
   DecoderRequest,
   CalldataDecoding,
@@ -64,7 +67,7 @@ export function* decodeCalldata(
   const contractType = Contexts.Utils.contextToType(context);
   isConstructor = context.isConstructor;
   const allocations = info.allocations.calldata;
-  let allocation: Allocation.CalldataAllocation;
+  let allocation: CalldataAllocation;
   let selector: string;
   //first: is this a creation call?
   if (isConstructor) {
@@ -186,9 +189,9 @@ export function* decodeEvent(
   let rawSelector: Uint8Array;
   let selector: string;
   let contractAllocations: {
-    [contextHash: string]: Allocation.EventAllocation;
+    [contextHash: string]: EventAllocation;
   }; //for non-anonymous events
-  let libraryAllocations: { [contextHash: string]: Allocation.EventAllocation }; //similar
+  let libraryAllocations: { [contextHash: string]: EventAllocation }; //similar
   const topicsCount = info.state.eventtopics.length;
   //yeah, it's not great to read directly from the state like this (bypassing read), but what are you gonna do?
   if (topicsCount > 0) {
@@ -228,8 +231,8 @@ export function* decodeEvent(
     info.contexts,
     codeAsHex
   );
-  let possibleContractAllocations: Allocation.EventAllocation[]; //excludes anonymous events
-  let possibleContractAnonymousAllocations: Allocation.EventAllocation[];
+  let possibleContractAllocations: EventAllocation[]; //excludes anonymous events
+  let possibleContractAnonymousAllocations: EventAllocation[];
   if (contractContext) {
     //if we found the contract, maybe it's from that contract
     const contextHash = contractContext.context;
